#769 add nested testsuite rule to briefwahl service
